feat: refactor CEP field component and update imports
